Fix EIP4844 spec types + Capella delegate types

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/DelegatingSpecConfigCapella.java
@@ -42,12 +42,12 @@ public class DelegatingSpecConfigCapella extends DelegatingSpecConfigBellatrix
   }
 
   @Override
-  public int getMaxWithdrawalsPerPayload() {
+  public UInt64 getMaxWithdrawalsPerPayload() {
     return specConfigCapella.getMaxWithdrawalsPerPayload();
   }
 
   @Override
-  public int getMaxBlsToExecutionChanges() {
+  public UInt64 getMaxBlsToExecutionChanges() {
     return specConfigCapella.getMaxBlsToExecutionChanges();
   }
 }
